[REFACTOR] 대기열 토큰 web interfaces 리팩토링
