+++ src/main/java/kr/hhplus/be/server/interfaces/queuetoken/controller/QueueTokenController.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domainold.queuetoken.web;
+package kr.hhplus.be.server.interfaces.queuetoken.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -6,11 +6,10 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import kr.hhplus.be.server.domainold.queuetoken.domain.QueueTokenService;
-import kr.hhplus.be.server.domainold.queuetoken.domain.dto.QueueTokenResponse;
-import kr.hhplus.be.server.domainold.queuetoken.web.dto.TokenRequest;
-import kr.hhplus.be.server.domainold.user.domain.User;
-import kr.hhplus.be.server.domainold.user.domain.UserService;
+import jakarta.validation.Valid;
+import kr.hhplus.be.server.domain.queuetoken.application.result.QueueTokenResult;
+import kr.hhplus.be.server.interfaces.queuetoken.dto.TokenRequest;
+import kr.hhplus.be.server.interfaces.usecase.QueueTokenCreateUseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
@@ -23,8 +22,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class QueueTokenController {
 
-    private final QueueTokenService queueTokenService;
-    private final UserService userService;
+    private final QueueTokenCreateUseCase queueTokenCreateUseCase;
 
     @Operation(
         summary = "대기열 토큰 발급",
@@ -32,7 +30,7 @@ public class QueueTokenController {
     )
     @ApiResponses(value = {
         @ApiResponse(responseCode = "200", description = "큐 토큰이 성공적으로 생성됨",
-            content = @Content(mediaType = "application/json", schema = @Schema(implementation = QueueTokenResponse.class))),
+            content = @Content(mediaType = "application/json", schema = @Schema(implementation = QueueTokenResult.class))),
         @ApiResponse(responseCode = "400", description = "잘못된 요청 데이터",
             content = @Content(mediaType = "application/json")),
         @ApiResponse(responseCode = "404", description = "사용자를 찾을 수 없음",
@@ -41,11 +39,9 @@ public class QueueTokenController {
             content = @Content(mediaType = "application/json"))
     })
     @PostMapping("/api/queue/tokens")
-    public ResponseEntity<QueueTokenResponse> createToken(
-        @RequestBody TokenRequest request) {
-        User user = userService.retrieveUser(request.getUserId());
+    public ResponseEntity<QueueTokenResult> createToken(@Valid @RequestBody TokenRequest request) {
 
-        QueueTokenResponse response = queueTokenService.create(user);
+        QueueTokenResult response = queueTokenCreateUseCase.create(request.getUserId());
 
         HttpHeaders headers = new HttpHeaders();
         headers.set("x-queue-token", response.getTokenUuid());
